08_user board 1:n model relation

+++ boards/views.py
@@ -24,7 +24,13 @@ def create(request):
             # content = form.cleaned_data.get('content')
             # 검증을 통과한 꺠긋한 데이터를 form 에서 가져와서 board를 만든다.
             # board = Board.objects.create(title=title, content=content)
-            board = form.save()
+            
+            # board를 바로 저장하지 않고 현재 user를 넣고 저장하기 위해 커밋=false를 씀.
+            # 실제 DB에  반영 전까지 단꼐를 진행하고 그 중간에 user 정보를
+            # request.user에서 가져와서 그 후에 저장한다.
+            board = form.save(commit=False)
+            board.user = request.user
+            board.save()
             return redirect('boards:detail', board.pk)
     # GET 요청(호은 다른 메서드)이면 기본 폼을 생선한다.
     else:
@@ -42,29 +48,35 @@ def detail(request, board_pk):
     
 def delete(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        board.delete()
-        return redirect('boards:index')
+    if board.user == request.user:
+        if request.method == 'POST':
+            board.delete()
+            return redirect('boards:index')
+        else:
+            return redirect('boards:detail', board.pk)
     else:
-        return redirect('boards:detail', board.pk)
-    
+        return redirect('boards:index')
+        
 @login_required    
 def update(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        form = BoardForm(request.POST, instance=board)
-        if form.is_valid():
-            # board.title = form.cleaned_data.get('title')
-            # board.content = form.cleaned_data.get('content')
-            # board.save()
-            board = form.save()
-            return redirect('boards:detail', board.pk)
-    # GET 요청이면(수정하기 버튼을 눌렀을 때)
+    if board.user == request.user:
+        if request.method == 'POST':
+            form = BoardForm(request.POST, instance=board)
+            if form.is_valid():
+                # board.title = form.cleaned_data.get('title')
+                # board.content = form.cleaned_data.get('content')
+                # board.save()
+                board = form.save()
+                return redirect('boards:detail', board.pk)
+        # GET 요청이면(수정하기 버튼을 눌렀을 때)
+        else:
+            # BoardForm 을 초기화(사용자 입력 값을 넣어준 상태로)
+            # form = BoardForm(initial={'title': board.title, 'content': board.content)
+            # form = BoardForm(initial=board.__dict__)
+            form = BoardForm(instance=board)
     else:
-        # BoardForm 을 초기화(사용자 입력 값을 넣어준 상태로)
-        # form = BoardForm(initial={'title': board.title, 'content': board.content)
-        # form = BoardForm(initial=board.__dict__)
-        form = BoardForm(instance=board)
+        return redirect('boards:index')
     # 1. POST: 요처에서 검증에 실패하였을때, 오류 메세지가 포함된 상태
     # 2. GET: 요청에서 초기화된 상태
     context = {'form': form,
